refactor(pms): route wms shared inbound reads through integration client
