+++ app/wms/inbound/repos/item_lookup_repo.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.pms.export.items.contracts.item_policy import ItemPolicy
-from app.pms.export.items.services.item_read_service import ItemReadService
+from app.integrations.pms.contracts import ItemPolicy
+from app.integrations.pms.inprocess_client import InProcessPmsReadClient
 
 
 async def get_item_policy_by_id(
@@ -11,8 +11,7 @@ async def get_item_policy_by_id(
     *,
     item_id: int,
 ) -> ItemPolicy | None:
-    svc = ItemReadService(session)
-    return await svc.aget_policy_by_id(item_id=int(item_id))
+    return await InProcessPmsReadClient(session).get_item_policy(item_id=int(item_id))
 
 
 __all__ = ["get_item_policy_by_id"]
