+++ app/wms/inbound/services/inbound_commit_service.py
@@ -8,8 +8,7 @@ from uuid import uuid4
 from fastapi import HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.pms.export.items.services.item_read_service import ItemReadService
-from app.pms.export.uoms.services.uom_read_service import PmsExportUomReadService
+from app.integrations.pms.inprocess_client import InProcessPmsReadClient
 from app.procurement.services.purchase_order_completion_sync import (
     sync_purchase_completion_for_inbound_event,
 )
@@ -76,7 +75,7 @@ async def _require_ratio_to_base(
     item_id: int,
     uom_id: int,
 ) -> int:
-    uom = await PmsExportUomReadService(session).aget_by_id(item_uom_id=int(uom_id))
+    uom = await InProcessPmsReadClient(session).get_uom(item_uom_id=int(uom_id))
     if uom is None or int(uom.item_id) != int(item_id):
         raise HTTPException(
             status_code=400,
@@ -95,8 +94,9 @@ async def _load_item_display_snapshot(
     item_id: int,
     uom_id: int,
 ) -> tuple[str | None, str | None, str | None]:
-    item = await ItemReadService(session).aget_basic_by_id(item_id=int(item_id))
-    uom = await PmsExportUomReadService(session).aget_by_id(item_uom_id=int(uom_id))
+    pms_client = InProcessPmsReadClient(session)
+    item = await pms_client.get_item_basic(item_id=int(item_id))
+    uom = await pms_client.get_uom(item_uom_id=int(uom_id))
 
     if item is None or uom is None or int(uom.item_id) != int(item_id):
         raise HTTPException(
